approve: Factor out findApprovers
As a step towards recycling this logic for other purposes
(e.g. merge-time commit tags [1]).

Stuffing hasApprovedLabel and notifications into the state struct
makes the provenance of the state data a bit less clear.  But I've
attempted to mitigate that with a comment in the state struct, and it
didn't seem to be worth defining a new structure for the things
findApprovers calculated that needed to be passed back to handle but
which didn't fit in approvers.Approvers.

I've also shifted the cancel docs up into the comment-iteration
section, since that's how the code handles cancels.

[1]: https://github.com/kubernetes/test-infra/pull/9400#discussion_r218209369