prow/flagutil/config: Add ConfigOptions
To centralize handling for -config-path and -job-config-path.

The refresh command previously had no -job-config-path and hard-coded
an empty string.  We could add a ConfigOptions.SkipJobConfigPath
boolean to turn off the flag injection for it, but I don't think it's
worth it.  For example, see 42060f7f (docs for gerrit controller,
2018-09-27, #9603), which recently added -job-config-path to the
gerrit command.

Several of the commands (ci-janitor, branchprotector, mkpj, and
peribolos) used to load, but not start, the config agent.  I'd
initially set this up as:

  func (o *ConfigOptions) Agent(start bool) (agent *config.Agent, err error) {

so callers could select whether to start the agent or not, but Steve
prefers the simpler API and just starting the agent every time.
Commands that don't need a running agent won't suffer greatly from the
additional resource consumption.

[1]: https://github.com/kubernetes/test-infra/pull/9701#discussion_r223294135